moved interfaces to types
