+++ src/client/datascience/serviceRegistry.ts
@@ -39,11 +39,11 @@ import { ExportBase } from './export/exportBase';
 import { ExportManager } from './export/exportManager';
 import { ExportManagerDependencyChecker } from './export/exportManagerDependencyChecker';
 import { ExportManagerFileOpener } from './export/exportManagerFileOpener';
-import { ExportManagerFilePicker, IExportManagerFilePicker } from './export/exportManagerFilePicker';
+import { ExportManagerFilePicker } from './export/exportManagerFilePicker';
 import { ExportToHTML } from './export/exportToHTML';
 import { ExportToPDF } from './export/exportToPDF';
 import { ExportToPython } from './export/exportToPython';
-import { ExportFormat, IExport, IExportManager } from './export/types';
+import { ExportFormat, IExport, IExportManager, IExportManagerFilePicker } from './export/types';
 import { GatherListener } from './gather/gatherListener';
 import { GatherLogger } from './gather/gatherLogger';
 import { DebugListener } from './interactive-common/debugListener';
